Add header for end game status and reset button
The message for the end game results is moved from the console log to
a special section on the board so that the users can see the end game
status. The reset button is added so that users can manually reset the
game without having to play it all the way out.